web/requirements.txt
Flask
Flask-WTF
Jinja2
Flask-DebugToolbar
ConfigParser
speedtest-cli
psutil